refactor(classfile): move ClassFileWritable to write.rs
