+++ javac-rs-classfile/src/frame.rs
@@ -1,7 +1,8 @@
-use crate::class::{ClassfileWritable, Tagged};
+use crate::class::Tagged;
 use crate::constpool::{ConstClassInfo, ConstPoolIndex};
 use crate::vec::JvmVecU2;
 use std::io::Write;
+use crate::writer::ClassfileWritable;
 
 #[derive(Eq, PartialEq, Debug)]
 pub enum StackMapFrame {
