+++ javac-rs-classfile/src/method.rs
@@ -3,7 +3,7 @@
 use crate::classfile_writable_mask_flags;
 
 use crate::attribute::AttributeInfo;
-use crate::class::ClassfileWritable;
+use crate::writer::ClassfileWritable;
 use crate::constpool::{ConstPoolIndex, ConstUtf8Info};
 use std::io::Write;
